+++ package/src/db/helpers.ts
@@ -54,7 +54,7 @@ export function excluded<T extends Table>(table: T): { [K in keyof T["_"]["colum
     const cols = getTableColumns(table);
     const tableName = getTableName(table);
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return new Proxy({} as any, {
+    return new Proxy(cols, {
         get(_, prop: string) {
             const col = cols[prop];
             if (!col) {